refactor: 잘못 부여된 Cannon 의 역할을 Rook 으로 변경
