+++ src/main/java/janggi/domain/piece/Cannon.java
@@ -1,19 +1,4 @@
 package janggi.domain.piece;
 
-import janggi.domain.Position;
-import janggi.domain.Side;
-
-public class Cannon extends Piece {
-
-    public Cannon(Side side, Position position) {
-        super(side, position);
-    }
-
-    @Override
-    protected boolean isMoveable(int x, int y) {
-        if (getPosition().hasSameX(x)) {
-            return !getPosition().hasSameY(y);
-        }
-        return getPosition().hasSameY(y);
-    }
+public class Cannon {
 }
