+++ src/main/java/janggi/domain/piece/Rook.java
@@ -1,4 +1,19 @@
 package janggi.domain.piece;
 
-public class Rook {
+import janggi.domain.Position;
+import janggi.domain.Side;
+
+public class Rook extends Piece {
+
+    public Rook(Side side, Position position) {
+        super(side, position);
+    }
+
+    @Override
+    protected boolean isMoveable(int x, int y) {
+        if (getPosition().hasSameX(x)) {
+            return !getPosition().hasSameY(y);
+        }
+        return getPosition().hasSameY(y);
+    }
 }
